move corosyinc code in dedicated package
